requirements.txt
pandas==2.2.3
spotipy==2.25.0
python-dotenv==1.0.1
fastapi==0.115.8
streamlit==1.42.0